third_party/blink/web_tests/external/wpt/css/css-shapes/shape-outside/values/shape-outside-circle-005-expected.txt

This is a testharness.js-based test.
FAIL test unit: cm - circle(50cm) - inline assert_equals: expected "circle(50cm at 50% 50%)" but got "circle(50cm)"
FAIL test unit: mm - circle(50mm) - inline assert_equals: expected "circle(50mm at 50% 50%)" but got "circle(50mm)"
FAIL test unit: in - circle(50in) - inline assert_equals: expected "circle(50in at 50% 50%)" but got "circle(50in)"
FAIL test unit: pt - circle(50pt) - inline assert_equals: expected "circle(50pt at 50% 50%)" but got "circle(50pt)"
FAIL test unit: pc - circle(50pc) - inline assert_equals: expected "circle(50pc at 50% 50%)" but got "circle(50pc)"
FAIL test unit: em - circle(50em) - inline assert_equals: expected "circle(50em at 50% 50%)" but got "circle(50em)"
FAIL test unit: ex - circle(50ex) - inline assert_equals: expected "circle(50ex at 50% 50%)" but got "circle(50ex)"
FAIL test unit: ch - circle(50ch) - inline assert_equals: expected "circle(50ch at 50% 50%)" but got "circle(50ch)"
FAIL test unit: rem - circle(50rem) - inline assert_equals: expected "circle(50rem at 50% 50%)" but got "circle(50rem)"
FAIL test unit: vw - circle(50vw) - inline assert_equals: expected "circle(50vw at 50% 50%)" but got "circle(50vw)"
FAIL test unit: vh - circle(50vh) - inline assert_equals: expected "circle(50vh at 50% 50%)" but got "circle(50vh)"
FAIL test unit: vmin - circle(50vmin) - inline assert_equals: expected "circle(50vmin at 50% 50%)" but got "circle(50vmin)"
FAIL test unit: vmax - circle(50vmax) - inline assert_equals: expected "circle(50vmax at 50% 50%)" but got "circle(50vmax)"
FAIL test unit: cm - circle(50cm) - computed assert_equals: expected "circle(1889.76px at 50% 50%)" but got "circle(1889.76px)"
FAIL test unit: mm - circle(50mm) - computed assert_equals: expected "circle(188.976px at 50% 50%)" but got "circle(188.976px)"
FAIL test unit: in - circle(50in) - computed assert_equals: expected "circle(4800px at 50% 50%)" but got "circle(4800px)"
FAIL test unit: pt - circle(50pt) - computed assert_equals: expected "circle(66.6667px at 50% 50%)" but got "circle(66.6667px)"
FAIL test unit: pc - circle(50pc) - computed assert_equals: expected "circle(800px at 50% 50%)" but got "circle(800px)"
FAIL test unit: em - circle(50em) - computed assert_equals: expected "circle(800px at 50% 50%)" but got "circle(800px)"
FAIL test unit: ex - circle(50ex) - computed assert_equals: expected "circle(640px at 50% 50%)" but got "circle(640px)"
FAIL test unit: ch - circle(50ch) - computed assert_equals: expected "circle(800px at 50% 50%)" but got "circle(800px)"
FAIL test unit: rem - circle(50rem) - computed assert_equals: expected "circle(800px at 50% 50%)" but got "circle(800px)"
FAIL test unit: vw - circle(50vw) - computed assert_equals: expected "circle(400px at 50% 50%)" but got "circle(400px)"
FAIL test unit: vh - circle(50vh) - computed assert_equals: expected "circle(300px at 50% 50%)" but got "circle(300px)"
FAIL test unit: vmin - circle(50vmin) - computed assert_equals: expected "circle(300px at 50% 50%)" but got "circle(300px)"
FAIL test unit: vmax - circle(50vmax) - computed assert_equals: expected "circle(400px at 50% 50%)" but got "circle(400px)"
Harness: the test ran to completion.
